Fixed typo in unit test

+++ Tests/SurgeBenchmarkTests/ArithmeticTests.swift
@@ -99,7 +99,7 @@ class ArithmeticTests: XCTestCase {
     // MARK: - Modulo: In Place
 
     func test_mod_in_place_array_array_float() {
-        measure_inout_array_array(of: Double.self) { measure in
+        measure_inout_array_array(of: Float.self) { measure in
             measureMetrics([.wallClockTime], automaticallyStartMeasuring: false) {
                 measure(Surge.modInPlace)
             }
